[rename] File 관련 클래스들 패키지 구조 개편
