+++ src/main/java/com/prgrms/rg/domain/common/file/application/FileIOException.java
@@ -1,4 +1,4 @@
-package com.prgrms.rg.domain.common.file;
+package com.prgrms.rg.domain.common.file.application;
 
 public class FileIOException extends RuntimeException {
 	public FileIOException(String message, Throwable cause) {
