+++ src/main/java/com/prgrms/rg/domain/common/file/model/FileRepository.java
@@ -1,4 +1,6 @@
-package com.prgrms.rg.domain.common.file;
+package com.prgrms.rg.domain.common.file.model;
+
+import com.prgrms.rg.domain.common.file.model.StoredFile;
 
 public interface FileRepository {
 	StoredFile save(StoredFile storedFile);
