+++ src/main/java/com/prgrms/rg/infrastructure/file/FileDeleteEntityListener.java
@@ -1,8 +1,11 @@
-package com.prgrms.rg.domain.common.file;
+package com.prgrms.rg.infrastructure.file;
 
 import javax.persistence.PreRemove;
 
 import org.springframework.beans.factory.annotation.Autowired;
+
+import com.prgrms.rg.domain.common.file.application.FileStore;
+import com.prgrms.rg.domain.common.file.model.StoredFile;
 
 public class FileDeleteEntityListener {
